remove internal types importing from public types
tried to break them up into each categories

also added helper.d.ts to try avoid circular
dependencies between each files imports
